fix linter error
