+++ lib/player.rb
@@ -1,0 +1,10 @@
+module TicTacToe
+  class Player
+    attr_reader :player1, :player2
+
+    def initialize(player1, player2)
+      @player1 = player1
+      @player2 = player2
+    end
+  end
+end
